// src/app/models/report/reportDepartmentIncomeDetailData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ReportDepartmentIncomeDetailData {

    courseCode?: string;
	coursepublicId?: number;
	publicNameTh?: string;
	publicNameEn?: string;
	remainAmount?: number;
	costShareDepAmount?: number;
	costShareDepPercent?: number;
	costShareGlobalAmount?: number;
	costShareGlobalPercent?: number;
	costShareCenterAmount?: number;
	costShareCenterPercent?: number;
	costShareManageAmount?: number;
	costShareManagePercent?: number;
	subjectCodeTh?: string;
	subjectCodeEn?: string;
	depIdLevel1?: number;
    rowNum?: number;
    size?: number;
    first?: number;
    mode?: string;

    registerDateStart?: Date;
    registerDateEnd?: Date;
    depId?: number;
    coordinatorTh?: string;

	depNameTh?: string;
	depNameEn?: string;

    maxEnroll?: number;
	tranCount?: number;

}